requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Agent Orchestration (Phase 4-5)
langgraph==0.2.45
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1
asyncpg==0.29.0

# Speech-to-Text (Phase 3)
faster-whisper>=1.0.0  # Requires FFmpeg - already installed for JCB-15

# Twitch Integration
twitchAPI==4.1.0
streamlink>=6.0.0  # For getting Twitch audio-only stream URLs

# OpenAI
openai==1.59.8

# Redis
redis==5.0.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# Environment & Config
python-dotenv==1.0.0

# Video Embeddings (CLIP)
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
httpx==0.26.0
asyncio==3.4.3
websockets==13.1  # EventSub WebSocket client
boto3==1.35.73

# Testing
respx==0.21.0  # HTTP mocking for httpx
responses==0.25.0  # HTTP mocking for requests

